fix(CreatePineconeIndexModal): fix typo in modal description refactor(settings): move SettingTabContainer to settingTab.tsx and remove the component feat(settings): add index creation callback to refresh index list
